\u27a4 env OMP_NUM_THREADS=1 ./multi_thread_op_test 
[==========] Running 5 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 5 tests from OpMultiThread
[ RUN      ] OpMultiThread.mul
mul speed up ratio:
	1	1
	2	2.25852
	3	3.36458
	4	4.46256
	5	5.55798
	6	5.79593
[       OK ] OpMultiThread.mul (893 ms)
[ RUN      ] OpMultiThread.batch_norm_channel_3
batch_norm: ChannelSize 3, is_test1
batch_norm speed up ratio:
	1	1
	2	1.97563
	3	2.88146
	4	3.76476
	5	4.41138
	6	4.79289
[       OK ] OpMultiThread.batch_norm_channel_3 (716 ms)
[ RUN      ] OpMultiThread.batch_norm_channel_10
batch_norm: ChannelSize 10, is_test1
batch_norm speed up ratio:
	1	1
	2	1.76583
	3	1.71796
	4	1.94259
	5	1.79704
	6	1.43413
[       OK ] OpMultiThread.batch_norm_channel_10 (3750 ms)
[ RUN      ] OpMultiThread.batch_norm_channel_20
batch_norm: ChannelSize 20, is_test1
batch_norm speed up ratio:
	1	1
	2	0.803463
	3	0.848495
	4	0.660625
	5	0.575842
	6	0.450057
[       OK ] OpMultiThread.batch_norm_channel_20 (23762 ms)
[ RUN      ] OpMultiThread.batch_norm_channel_30
batch_norm: ChannelSize 30, is_test1
batch_norm speed up ratio:
	1	1
	2	0.87689
	3	0.67897
	4	0.570793
	5	0.513415
	6	0.467967
[       OK ] OpMultiThread.batch_norm_channel_30 (43440 ms)
[----------] 5 tests from OpMultiThread (72561 ms total)

[----------] Global test environment tear-down
[==========] 5 tests from 1 test case ran. (72561 ms total)
[  PASSED  ] 5 tests.