@tensor-tang 2018-08-08
我试了下，能复现这个结果。

同时，先不看结果的正确性，如果把https://github.com/PaddlePaddle/Paddle/blob/7c8b69c700a19d777f27ae0495cd2129d3156a9c/paddle/fluid/operators/batch_norm_op.cc#L289-L318
 这段注释掉，性能比例可以基本正常

```
44: [       OK ] OpMultiThread.batch_norm_channel_10 (4692 ms)
44: [ RUN      ] OpMultiThread.batch_norm_channel_20
44: batch_norm: ChannelSize 20, is_test1
44: batch_norm speed up ratio:
44: 	1	1
44: 	2	1.96919
44: 	3	2.96584
44: 	4	3.90309
44: 	5	4.84115
44: 	6	5.82826
44: [       OK ] OpMultiThread.batch_norm_channel_20 (4761 ms)
44: [ RUN      ] OpMultiThread.batch_norm_channel_30
44: batch_norm: ChannelSize 30, is_test1
44: batch_norm speed up ratio:
44: 	1	1
44: 	2	1.97696
44: 	3	2.93142
44: 	4	3.8971
44: 	5	4.86192
44: 	6	5.85003
```